r10606/sofa : push b69272b..14b982c
* commit 14b982c28444a55d190077d27d56004b24d2d5b2
  Author: Marc Legendre <marc.legendre@inria.fr>
  Date:   Mon Feb 3 12:03:41 2014 +0100
  
      Fix some mismatches between integer types
  
   extlibs/qwt-6.0.1/src/qwt_series_data.h            |    2 +-
   framework/sofa/core/behavior/MechanicalState.h     |    4 ++--
   framework/sofa/core/collision/DetectionOutput.h    |    2 +-
   framework/sofa/core/topology/BaseMeshTopology.cpp  |    2 +-
   framework/sofa/core/topology/BaseMeshTopology.h    |   14 +++++++-------
   .../sofa/core/topology/TopologyElementHandler.h    |    2 +-
   framework/sofa/defaulttype/VecTypes.h              |   10 +++++-----
   framework/sofa/helper/fixed_array.h                |    2 +-
   framework/sofa/helper/gl/glfont.cpp                |    2 +-
   framework/sofa/helper/io/MeshOBJ.cpp               |    2 +-
   framework/sofa/helper/io/MeshSTL.cpp               |    4 ++--
   framework/sofa/helper/io/MeshTopologyLoader.cpp    |    4 ++--
   framework/sofa/helper/vector.h                     |    6 ++++--
   modules/sofa/simulation/common/MechanicalVisitor.h |    4 ++--
   modules/sofa/simulation/common/SceneLoaderPHP.cpp  |    2 +-
   15 files changed, 32 insertions(+), 30 deletions(-)



Former-commit-id: 5dd02fe7920bfa9311ac954d3cd883b48577822e

+++ extlibs/qwt-6.0.1/src/qwt_series_data.h
@@ -257,7 +257,7 @@ size_t QwtArraySeriesData<T>::size() const
 template <typename T>
 T QwtArraySeriesData<T>::sample( size_t i ) const
 {
-    return (T)d_samples[i];
+    return (T)d_samples[(int)i];
 }
 
 //! Interface for iterating over an array of points
